renames conduit package
