+++ network/p2p/conduit/conduit.go
@@ -1,4 +1,4 @@
-package p2p
+package conduit
 
 import (
 	"context"
